requirements.txt
blinker==1.6.2
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pdfminer.six==20221105
pycparser==2.21
Werkzeug==2.3.7
